refactor: move config migrations to own file
